Squashed commit of the following:
commit 0fb2b592704bbd704823b526f3c968a23ef90f8a
Merge: 55157b6 59a3458
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Thu Jun 11 23:52:28 2020 +0000

    Merge branch 'save-restapi-todb' into 'develop'

    save rest api operation to db.

    See merge request gcoin/gfashion/gfashion-java-proxy!107

commit 55157b6ecf1ed228af2110cb12257d85ba1168c5
Merge: dac9610 b13c5b7
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Thu Jun 11 23:51:53 2020 +0000

    Merge branch 'update-product-key' into 'develop'

    using 'id' as primary key.

    See merge request gcoin/gfashion/gfashion-java-proxy!106

commit 59a34583ea4ffc14fe8b61f6389eab264fb6e696
Author: eric <ericback2019@protonmail.com>
Date:   Thu Jun 11 16:01:02 2020 +0800

    save rest operation to db.

commit b13c5b7cd270bd01649d581dc81046043f4cdca5
Author: eric <ericback2019@protonmail.com>
Date:   Thu Jun 11 12:04:02 2020 +0800

    using 'id' as primary key.

commit dac9610c123deb6fb09cc14fd5ea63f101e01138
Merge: 6061c39 45233b7
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Thu Jun 11 11:08:39 2020 +0000

    Merge branch 'feature-gf123-search' into 'develop'

    Feature gf123 search

    See merge request gcoin/gfashion/gfashion-java-proxy!96

commit 45233b7479e50f0fc77973fab6818f69deb04503
Author: angus <neocyber35@gmail.com>
Date:   Thu Jun 11 11:08:39 2020 +0000

    Feature gf123 search

commit 6061c3917aac48f448c3a55d2d1b02ee43f1c505
Merge: b442547 561a915
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Thu Jun 11 09:39:49 2020 +0000

    Merge branch 'feature-remove-it-ignore' into 'develop'

    刪除IT的ignore標示

    See merge request gcoin/gfashion/gfashion-java-proxy!105

commit 561a915527071ec7132959257e7e722302678878
Author: Mabea Li <mabeali@test.com>
Date:   Thu Jun 11 18:16:16 2020 +0900

    刪除IT的ignore標示

commit b44254716018e8b3d890841477f04140bee9b17b
Merge: dd8c8b9 1fdf035
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Thu Jun 11 07:01:55 2020 +0000

    Merge branch 'feature-bigdecimal' into 'develop'

    添加購物車商品價格BigDecimal相關的測試

    See merge request gcoin/gfashion/gfashion-java-proxy!100

commit 1fdf0355f3fd2e55ada79cc117eae31c0be11cb5
Author: Mabea Li <mabeali@test.com>
Date:   Thu Jun 11 15:54:50 2020 +0900

    Fix IT fail

commit dd8c8b9decab9e907f907c0fe0bff5df8f4e1636
Author: Grant(gznz) <gznz@protonmail.com>
Date:   Thu Jun 11 18:47:09 2020 +1200

    Revert "revert gitlab-ci test stage" and ignore failed Cart*IT

    This reverts commit a52a911ff17c3231ee1f9008abd743043b8e32bd.

    # Conflicts:
    #	.gitlab-ci.yml

commit f4cc1b5343e7cb839f8943cc7fac2de6ba726d76
Merge: eeead07 95a335c
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Thu Jun 11 06:15:53 2020 +0000

    Merge branch 'fixITs' into 'develop'

    Fix  ITs

    See merge request gcoin/gfashion/gfashion-java-proxy!103

commit 95a335cdb4e3943f6592f63134c0e1d57a479e0d
Author: Andrew Luo <andrewluo.us@gmail.com>
Date:   Thu Jun 11 06:15:53 2020 +0000

    Fix  ITs

commit eeead0728fdfc7df0c305a4e7c41d95f8141d683
Merge: 87e8606 509e5ad
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Thu Jun 11 05:39:50 2020 +0000

    Merge branch 'fixITs' into 'develop'

    FixITs

    See merge request gcoin/gfashion/gfashion-java-proxy!102

commit 509e5ad9e57c9908216d6e627082ebe541c576a3
Author: andrewluo.us <andrewluo.us@gmail.com>
Date:   Wed Jun 10 21:47:53 2020 -0700

    FixITs

commit 753244c4067333a60bd030dd2ed946a3a341f50f
Author: Mabea Li <mabeali@test.com>
Date:   Wed Jun 10 21:43:42 2020 +0900

    添加購物車商品價格BigDecimal相關的測試

commit 87e8606c1ca8d27762d421563a4f460d4e4aa8f1
Author: Grant(gznz) <gznz@protonmail.com>
Date:   Thu Jun 11 15:38:31 2020 +1200

    work around gitlab-ci test stage issue

commit a52a911ff17c3231ee1f9008abd743043b8e32bd
Author: Grant(gznz) <gznz@protonmail.com>
Date:   Thu Jun 11 14:17:29 2020 +1200

    revert gitlab-ci test stage

commit ef957e83f124563e153d322445511777a300102a
Merge: a3d39e1 23a9ec4
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Thu Jun 11 01:37:55 2020 +0000

    Merge branch 'updateDesignerEntity' into 'develop'

    update designer entity

    See merge request gcoin/gfashion/gfashion-java-proxy!97

commit a3d39e1bb081ca08386edde849d6d772ada090be
Merge: e965349 7cd53d5
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Wed Jun 10 23:34:26 2020 +0000

    Merge branch 'improve-customer-api' into 'develop'

    Structuring Returned Categories

    See merge request gcoin/gfashion/gfashion-java-proxy!89

commit 7cd53d590139b3a54a98be129fc4ee2fd027da67
Author: Andrew Luo <andrewluo.us@gmail.com>
Date:   Wed Jun 10 23:34:26 2020 +0000

    Structuring Returned Categories

commit e9653491ed43aac5306986eb2162bb9789ea96b5
Merge: 7ce0642 03313c6
Author: bgping.guo <etnlona@gmail.com>
Date:   Wed Jun 10 15:36:57 2020 +0000

    Merge branch 'add-ddb-operation-log' into 'develop'

    add ddb operation log aop.

    See merge request gcoin/gfashion/gfashion-java-proxy!95

commit 03313c6dddd482f792101180ed707f442b7d872d
Author: eric <ericback2019@protonmail.com>
Date:   Wed Jun 10 23:07:15 2020 +0800

    add ddb operation log aop.

commit 7ce0642247d1306daffa70b17bb1e57a77d94504
Merge: 7eec3a9 2d36e69
Author: bgping.guo <etnlona@gmail.com>
Date:   Wed Jun 10 15:29:14 2020 +0000

    Merge branch 'using-magentoId-for-ddb-primary-key' into 'develop'

    chang product table primary key from dynamodb autogenkey to magento productId.

    See merge request gcoin/gfashion/gfashion-java-proxy!94

commit 2d36e69b35742d052f12863c20a5a249c3ee054c
Author: eric <ericback2019@protonmail.com>
Date:   Wed Jun 10 22:28:51 2020 +0800

    chang product table primary key from dynamodb autogenkey to magento productId.

commit 7eec3a965a63c24c1d0e2f47224c672afe268fdd
Merge: eab9b48 a1a9462
Author: bgping.guo <etnlona@gmail.com>
Date:   Wed Jun 10 15:12:35 2020 +0000

    Merge branch 'add_dynamodb_entity_order_shipment' into 'develop'

    Add dynamodb entity order shipment

    See merge request gcoin/gfashion/gfashion-java-proxy!91

commit a1a9462309be1c59c2b6d7689c0455e4dc21d02c
Author: curl123 <pp5066t@gmail.com>
Date:   Wed Jun 10 23:37:58 2020 +0900

    change product

commit ae40e9fbc24c3723baa1731ff2011a1ac0c1cb28
Author: curl123 <pp5066t@gmail.com>
Date:   Wed Jun 10 23:09:47 2020 +0900

    change method name

commit 23a9ec4013a7235927be9e32ab9da22e9f7ca6c1
Author: Candy_xd <serezhautliakov1999@gmail.com>
Date:   Wed Jun 10 22:09:09 2020 +0800

    update designer entity

commit 55b0d03a0c1b97c06fc2762aa5e4a1dc8fbed568
Author: curl123 <pp5066t@gmail.com>
Date:   Wed Jun 10 22:44:02 2020 +0900

    add dynamodb entity order shipment

commit 21550e083810335add62699f774fe4a3852af9bd
Author: curl123 <pp5066t@gmail.com>
Date:   Wed Jun 3 15:03:33 2020 +0900

    add comments

commit ab19e4de0bdda7b76e7c08f963c3f05c9927481f
Author: curl123 <pp5066t@gmail.com>
Date:   Tue Jun 2 21:53:11 2020 +0900

    add multiple language

commit 84e0cdc25073e7387de50c64caf0960529e9a22d
Author: curl123 <pp5066t@gmail.com>
Date:   Sat May 30 18:28:12 2020 +0900

    add_gfvendor

commit eab9b4851bfc4a883166b11e7076686def1087c1
Author: Grant(gznz) <gznz@protonmail.com>
Date:   Wed Jun 10 21:25:21 2020 +1200

    tweak gitlab-ci test stage for easier integration testing with Spring Boot and REST-Assured

commit 264c33a95561fc4034c95dca150264a368bf0c58
Merge: 4ad0634 50c2639
Author: gznz <254-gznz@users.noreply.gitlab.freevog.ml>
Date:   Wed Jun 10 07:07:52 2020 +0000

    Merge branch 'develop_split_build_test' into 'develop'

    Develop split build test

    See merge request gcoin/gfashion/gfashion-java-proxy!90

commit 50c26393c6d8255cdac36f968a1207903fc83186
Author: nail_albumen <4ruyobti1@protonmail.com>
Date:   Wed Jun 10 07:07:52 2020 +0000

    Develop split build test

+++ g-fashion-client/src/main/java/com/gfashion/restclient/magento/cart/MagentoCartNegotiableQuote.java
@@ -6,6 +6,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.json.JSONObject;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -21,7 +23,7 @@ public class MagentoCartNegotiableQuote {
     @SerializedName("negotiated_price_value")
     private Integer negotiatedPriceValue;
     @SerializedName("shipping_price")
-    private Integer shippingPrice;
+    private BigDecimal shippingPrice;
     @SerializedName("quote_name")
     private String quoteName;
     @SerializedName("expiration_period")
@@ -46,13 +48,13 @@ public class MagentoCartNegotiableQuote {
     @SerializedName("creatorType")
     private Integer creatorType;
     @SerializedName("original_total_price")
-    private Integer originalTotalPrice;
+    private BigDecimal originalTotalPrice;
     @SerializedName("base_original_total_price")
-    private Integer baseOriginalTotalPrice;
+    private BigDecimal baseOriginalTotalPrice;
     @SerializedName("negotiated_total_price")
-    private Integer negotiatedTotalPrice;
+    private BigDecimal negotiatedTotalPrice;
     @SerializedName("base_negotiated_total_price")
-    private Integer baseNegotiatedTotalPrice;
+    private BigDecimal baseNegotiatedTotalPrice;
     @SerializedName("extension_attributes")
     private JSONObject extensionAttributes;
 }
