+++ g-fashion-domain/src/main/java/com/gfashion/domain/cart/GfCartTotalSegment.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @Data
 @AllArgsConstructor
@@ -13,7 +15,7 @@ public class GfCartTotalSegment {
 
     private String code;
     private String title;
-    private Integer value;
+    private BigDecimal value;
     private GfCartTotalSegmentExtensionAttributes extensionAttributes;
     private String area;
 }
